fix(hal): converge the gcc-13 -Wconversion tail + keep-going UBSan builds
- ra8_fs exfat bitmap scan: promote sec[byte] explicitly before the
  shift so the int promotion never reaches the unsigned mask.
- clang-format-22 realignment + comment-format pass on the swept TUs.
- MISRA ratchet baseline regenerated (the 1U literals shifted a handful
  of essential-type buckets).
- tests/build_tests.sh builds with -k so a red gate reports every
  failing TU in one run instead of the first.
- The UBSan CI job pins CC=gcc-13 like the coverage job: ambient gcc
  changes with the runner image and -Wconversion findings are
  compiler-version-specific.

+++ libs/ra8_fs/src/ra8_fs_fat_exfat_write.c
@@ -127,7 +127,7 @@ priv_exfat_bitmap_scan(const ra8_fs_mount_t* m, uint32_t bmp_lba, uint32_t need,
     }
     const uint32_t byte = (idx >> k_exfat_bit_shift) % k_ra8_fs_bytes_per_sector;
     const uint32_t bit  = idx & k_exfat_bit_mask;
-    if (((sec[byte] >> bit) & 1U) != 0U) {
+    if ((((uint32_t)sec[byte] >> bit) & 1U) != 0U) {
       run = 0U;
       continue;
     }
